Added MenuModel class to folder Model. It contains tree of menu and methods for work with it: getMenuData; addNode. Added method getMenuData to DataProvider for getting data from server firstly and from MenuModel later. Added method createMenuModel to DataProvider for creating menu tree. Added method getEntireMenuDataWithResponseBlock to RemoteDataProvider for get json-file from server with all menu tree and parse it. Added parseEntireMenu to DataParser for parse all menu tree in json-file.

+++ LV-116.iOS.SS.Restaurant/LV-116.iOS.SS.Restaurant/Models/MenuItem.h
@@ -12,11 +12,11 @@
 
 -(instancetype)initWithId:(int)Id categoryId:(int) categoryId description:(NSString*)description name:(NSString*)name portions:(int)portions price:(float) price;
 
-@property int Id;
-@property int categoryId;
+@property long Id;
+@property long categoryId;
 @property (strong, nonatomic) NSString *description;
 @property (strong, nonatomic) NSString *name;
-@property int portions;
+@property long portions;
 @property float price;
 
 @end
